+++ src/components/TimelineSec/Timeline.jsx
@@ -1,0 +1,96 @@
+import React, { useState, useEffect, useRef } from 'react';
+import { timelinecontent } from '../../../config/content/Timelinecontent';
+import {
+  ProgressBar,
+  Progress,
+  Branch,
+  MobileBranch,
+  Box,
+  BoxHeading,
+  Duration,
+  Description,
+  TimelineHead,
+} from './styles';
+
+const Line = () => (
+  <img
+    src='https://res.cloudinary.com/dgjzygzgx/image/upload/v1705664338/Line_3_xzegkw.svg'
+    alt='Line'
+  />
+);
+const Ellipse = () => (
+  <div>
+    <img
+      src='https://res.cloudinary.com/dgjzygzgx/image/upload/v1705901250/Ellipse_20_nhbab9.svg'
+      alt='Ellipse'
+      className='absolute bottom-0 scale-150'
+    />
+  </div>
+);
+
+export default function Timeline() {
+  const [scrollHeight, setScrollHeight] = useState(0);
+  const targetRef = useRef(null);
+
+  const handleScroll = () => {
+    if (targetRef.current) {
+      const { top, bottom } = targetRef.current.getBoundingClientRect();
+      if (bottom > 0 && top < window.innerHeight) {
+        let scrolled;
+
+        if (top >= 0) {
+          scrolled = Math.min(100, Math.max(0, (1 - top / (bottom - top)) * 100 - 30));
+        } else {
+          scrolled = Math.min(
+            100,
+            Math.max(0, (1 - (bottom - window.innerHeight) / (bottom - top)) * 100 - 18),
+          );
+        }
+
+        setScrollHeight(scrolled);
+      }
+    }
+  };
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(() => {}, { threshold: 0 });
+
+    if (targetRef.current) {
+      observer.observe(targetRef.current);
+    }
+
+    window.addEventListener('scroll', handleScroll);
+
+    return () => {
+      observer.disconnect();
+      window.removeEventListener('scroll', handleScroll);
+    };
+  }, []);
+
+  return (
+    <>
+      <TimelineHead>{timelinecontent.title}</TimelineHead>
+      <div className='flex m-auto' ref={targetRef}>
+        <ProgressBar>
+          <Progress height={scrollHeight}>
+            <Ellipse />
+          </Progress>
+        </ProgressBar>
+        <div className='flex flex-col md:gap-[36px]'>
+          {timelinecontent.nodes.map((node) => (
+            <div key={node.serialid} className='flex items-center md:flex-row flex-col'>
+              <Branch />
+              <Box color={node.color}>
+                <BoxHeading color={node.color}>{node.title}</BoxHeading>
+                <Line />
+                <Duration>{node.duration}</Duration>
+                <Description>{node.description}</Description>
+              </Box>
+              <MobileBranch />
+            </div>
+          ))}
+        </div>
+      </div>
+    </>
+  );
+}
