+++ src/pages/playground.js
@@ -6,12 +6,11 @@ import Layout from '../components/shared/Layout';
 import AboutSection from '../components/AboutUs/About';
 import SponsorSection from '../components/Sponsor/Sponsor';
 import FAQ from '../components/Faq/Faq';
-import TimelineSection from '../components/TimelineSection/Timeline';
 import { Footer } from '../components/marginals';
 import Prizes from '../components/Prizes/Prizes';
 import PreviousStats from '../components/PrevStats/PreviousStats';
 import ParticleBackground from '../components/shared/Particle';
-import Timeline2 from '../components/TimelineSec/Timeline2';
+import Timeline from '../components/TimelineSec/Timeline';
 
 const Homepage = ({ location }) => {
   const isHome = location?.pathname === '/';
@@ -55,8 +54,7 @@ const Homepage = ({ location }) => {
           <PreviousStats />
           <AboutSection />
           <Prizes />
-          <Timeline2 />
-          <TimelineSection />
+          <Timeline />
           <SponsorSection />
           <FAQ />
         </Container>
